Added webpack bundler, and separated front js to modules.
