+++ package.json
@@ -6,14 +6,18 @@
   "dependencies": {
     "cors": "^2.8.5",
     "express": "^4.17.1",
+    "webpack": "^5.62.1",
+    "webpack-cli": "^4.9.1"
+  },
+  "devDependencies": {
     "morgan": "^1.10.0",
     "nodemon": "^2.0.14"
   },
-  "devDependencies": {},
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node ./src/server.js",
-    "dev": "nodemon ./src/server.js"
+    "dev": "nodemon ./src/server.js",
+    "build": "webpack"
   },
   "repository": {
     "type": "git",
