+++ src/front/index.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Phonebook</title>
     <link rel="stylesheet" href="./style.css">
-    <script src="./index.js" defer></script>
+    <script src="../../dist/bundled.js" defer></script>
     <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
 </head>
 <body>
